moving custom components to a new package and few modifications with password show/hide icon.
