+++ src/com/avk/coffer/components/CofferTextField.java
@@ -1,4 +1,4 @@
-package com.avk.coffer;
+package com.avk.coffer.components;
 
 import java.awt.Color;
 import java.awt.Dimension;
@@ -9,6 +9,9 @@ import javax.swing.JTextField;
 import javax.swing.SwingConstants;
 import javax.swing.border.CompoundBorder;
 import javax.swing.border.EmptyBorder;
+
+import com.avk.coffer.CofferReferences;
+import com.avk.coffer.CofferRoundBorder;
 
 @SuppressWarnings("serial")
 public class CofferTextField extends JTextField {
