+++ src/com/avk/coffer/components/listeners/CofferPasswordBlankEditListener.java
@@ -1,4 +1,4 @@
-package com.avk.coffer;
+package com.avk.coffer.components.listeners;
 
 public interface CofferPasswordBlankEditListener {
 	public void onValueChange();
