+++ src/com/avk/coffer/components/listeners/CofferToggleSwitchListener.java
@@ -1,4 +1,4 @@
-package com.avk.coffer;
+package com.avk.coffer.components.listeners;
 
 import java.util.EventListener;
